netlink: Set 'ETHTOOL_FLAG_LEGACY' for compatibility with legacy ioctl interface
Set the above mentioned flag in the ethtool netlink request header of
the various commands mapped to 'ethtool set' (e.g.,
'ETHTOOL_MSG_LINKMODES_SET').

The purpose of the flag is to indicate to the kernel to be compatible
with legacy ioctl interface. The current use case is to ensure that the
kernel will advertise all the supported link modes when autoneg is
enabled, but without specifying other parameters.

To prevent the kernel from complaining about unknown flags, the flag is
only set in the request header in case the kernel supports it. This is
achieved by using the recently introduced per-operation policy dump
infrastructure.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>