ethtool: Add transceiver module extended states
Add support for an extended state and two extended sub-states to
describe link issues related to transceiver modules.

Output example:

XXX: Add me

In case "CMIS transceiver module is not in ModuleReady state" and the
module is in ModuleFault state, it is possible to read the fault reason
from the EEPROM dump.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>